+++ CHANGES.md
@@ -4,10 +4,12 @@
 
 - Fix: Throw proper `Error` on undefined child content
 - Fix: Allow previously undocumented fragment creation to accept children
+- Update: Remove now removed `entities`, `notations`, `internalSubset`
+  properties
 - Linting (ESLint): As per latest ash-nazg
 - Travis: Adding builds 10, 12
 - Testing: Coverage
-- Testing: Use builtin body, $, $$, nbsp
+- Testing: Use builtin `body`, `$`, `$$`, `nbsp`
 - npm: Update devDeps, peerDep
 
 ## 0.47.0